aspinwall: initial commit
...Well, technically this should be the "GTK4 rewrite" commit,
but since I was really early into developing this, I figured I'd
nuke the whole commit history (of 2 whole commits!) and start anew.

Currently the project is in a *very basic* state; there's a launcher
capable of showing the time and some basic widgets. Still, it took
me a whole day to get this up and running, most of which was spent
fighting with Gtk.Template; it can be a tricky beast to tame. I'm
planning to write a blog post detailing everything I know about it
so far, just to spare others the misfortune of having to blindly
work with it (most tutorials skip out on how Gtk.Template.Child()
works, which can be a huge pain).