widgets: add widget loading mechanism
Currently this is not noticeable, as there's no way to actually
add/list available widgets yet; however, it appears to work based
on some print statements I added during development.

(Caveat: the widget loading/saving mechanism in the launcher might
be broken, as it hasn't properly been tested yet. I had to hastily
change it during development as, suprise suprise, you can't store
a class in a JSON dump. I moved to an ID system for widgets instead,
and that resulted in the functions you see. As much as I hate to
commit code that might not work, it's probably a better idea than
rolling up everything I've added so far into one commit.)