launcher: redesign widget header, minor cosmetic changes to widget chooser
Also fixes a bug where the overlay caused the scrollbar + right corner of
widgets to be inaccessible. Technically the ideal fix would be to move the
menu button to a separate overlay, but I'll hold off on doing that until
I figure out how to get the launcher to display correctly on vertical screens.