launcher: AspWidget -> LauncherWidget
Historically (read: veeeery early in development) the main widget class
was to be used to display widgets in the launcher. However, as time went
on, it was split into two classes: one for the launcher display, and one
for the widget data.

Keeping this launcher display widget named "AspWidget" would be confusing
for developers.