launcher: hide widget header by default
The header can now be brought up by long-pressing the widget. There's
also a settings button that's supposed to appear when hovering over
the widget, but there doesn't seem to be a way to check the hover state
of a GtkBox. Maybe it'll be possible once we switch to ListView...

Also, the "x" button now hides the header; the remove button was moved
to a separate button on the left side of the header.