utils: new Dimmable mixin
The Dimmable mixin can be added to automatically add dim() and undim()
functions to widgets. Widgets that use this mixing must add the .dimmable
style class to the object. (We can't add this in-code, not without calling
add_style_class in the functions at least, since when multiple object are
inherited only the first one's init function is called, and if we have to
force inheritants to run a separate function, we might as well make them
set the class themselves.)