launcher: app-chooser: only add/remove needed icons
Prior to this, every time the app chooser was launcher or a favorite
was added, the box would re-create the entire list from scratch, which
would result in a slight slowdown on app chooser load and a nasty
memory leak bug (which I think is actually a pygobject issue where
Gio.ListStore objects are not freed properly, but I'll need to reproduce
this out-of-code to confirm my suspicions).

Now, a smaller function runs that compares the old app list with the
current app list, and adds/removes items from the main store as needed.
Also, the favorites box now uses the same liststore as the main app box,
which allows for automatic addition/removal on app deletion and makes
setting up the "add to favorites" and "remove from favorites" buttons
a bit easier.