launcher: wallpaper: add option to use Aspinwall's wallpaper settings
In commit 8455db2 (launcher: use GNOME background), we switched from using
our own wallpaper config value to GNOME's background value. The idea behind
this was that this would allow us to integrate better with the rest of GNOME's
apps.

However, GNOME's background settings are rather limited. Additionally, systems
without GNOME components would fail to load the wallpaper if GNOME's settings
schema was not found (in fact, with how we handled this before this commit,
the launcher would not launch at all).

Thus, we'll switch back to using our own wallpaper variable, with a switch
to use GNOME's values for compatibility. This commit also adds a settings
toggle for switching this.