launcher: change mentions of wallpaper-scaling to wallpaper-style
Earlier in development, wallpaper-scaling was only meant to store
wallpaper *scaling* options, but I added the "solid color" option
as well. To keep things consistent with GNOME, I renamed the option
to "Wallpaper style" but forgot to change the actual option name.
This commit fixes that, and changes a few references to "wallpaper
scaling" in the settings code.